# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OrbitGuard API",
    "summary": "Constellation resilience studio for the КосмоХакатон 2026 case",
    "description": "Loads `cosmo-A-1.0` scenarios, simulates a constellation over its time grid, routes traffic from client sites to a gateway, and reports availability, outages and per-satellite criticality.\n\n**Conventions.** Times are seconds from the start of the run (`t_s`), angles are degrees, distances kilometres, fractions are 0…1 (not percent). Keys are `snake_case`, matching the official case format.",
    "version": "dev"
  },
  "paths": {
...
    "/api/v1/scenarios": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Import a scenario from a JSON body",
        "operationId": "import_scenario_api_v1_scenarios_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "overwrite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overwrite"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "List available scenarios",
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  },
                  "title": "Response List Scenarios Api V1 Scenarios Get"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Fetch one scenario with its original document",
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenarios"
        ],
        "summary": "Delete an imported scenario",
        "operationId": "delete_scenario_api_v1_scenarios__scenario_id__delete",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{scenario_id}/download": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Download the scenario as the original cosmo-A-1.0 file",
        "operationId": "download_scenario_api_v1_scenarios__scenario_id__download_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioDetail": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/ScenarioSummary"
          },
          "scenario": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenario"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "scenario"
        ],
        "title": "ScenarioDetail",
        "description": "Everything the configuration panel and globe need to draw an unrun scenario."
      },
      "ScenarioSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source": {
            "type": "string",
            "enum": [
              "official",
              "imported"
            ],
            "title": "Source"
          },
          "satellite_count": {
            "type": "integer",
            "title": "Satellite Count"
          },
          "plane_count": {
            "type": "integer",
            "title": "Plane Count"
          },
          "client_count": {
            "type": "integer",
            "title": "Client Count"
          },
          "gateway_count": {
            "type": "integer",
            "title": "Gateway Count"
          },
          "launch_stage": {
            "type": "integer",
            "title": "Launch Stage"
          },
          "horizon_s": {
            "type": "integer",
            "title": "Horizon S"
          },
          "step_s": {
            "type": "integer",
            "title": "Step S"
          },
          "steps": {
            "type": "integer",
            "title": "Steps"
          },
          "target_availability": {
            "type": "number",
            "title": "Target Availability"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "satellite_count",
          "plane_count",
          "client_count",
          "gateway_count",
          "launch_stage",
          "horizon_s",
          "step_s",
          "steps",
          "target_availability"
        ],
        "title": "ScenarioSummary",
        "description": "One row in the scenario picker."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}